backend/requirements.txt
addict==2.4.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
bcc==0.18.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.9.0
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.1.8
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
ConfigArgParse==1.7
confluent-kafka==2.8.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dash==3.0.2
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
ecdsa==0.19.1
evdev==1.4.0
exceptiongroup==1.2.2
executing==2.2.0
fail2ban==0.11.2
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.29.1
fs==2.4.12
fsspec==2024.6.1
gast==0.5.2
h11==0.14.0
hidpidaemon==18.4.6
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipython==8.35.0
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
kafka-python==2.0.6
kazam==1.4.5
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
MarkupSafe==2.1.5
matplotlib==3.5.1
matplotlib-inline==0.1.7
more-itertools==8.10.0
mpmath==1.3.0
narwhals==1.34.1
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.3
numpy==1.24.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
open3d==0.19.0
packaging==21.3
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
Pivy==0.6.5
platformdirs==4.3.7
plotly==6.0.1
ply==3.11
pop-transition==1.1.2
prompt_toolkit==3.0.50
protobuf==3.12.4
protontricks==1.7.0
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pycairo==1.20.1
pycups==2.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pydbus==0.6.0
Pygments==2.19.1
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyquaternion==0.9.9
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-jose==3.4.0
python-magic==0.4.24
python-multipart==0.0.20
python-xlib==0.29
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qrcode==8.1
redis==5.2.1
referencing==0.36.2
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
retrying==1.3.4
rpds-py==0.24.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
setproctitle==1.2.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
systemd-python==234
threadpoolctl==3.6.0
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unicodedata2==14.0.0
urllib3==1.26.5
uvicorn==0.34.0
vdf==3.4
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.0.6
widgetsnbextension==4.0.14
xdg==5
xkit==0.0.0
zipp==1.0.0
